still int

+++ opendbc/safety/tests/common.py
@@ -166,8 +166,8 @@ class LongitudinalGasBrakeSafetyTest(PandaSafetyTestBase, abc.ABC):
   MIN_GAS: int = 0
   MAX_GAS: int | None = None
   INACTIVE_GAS = 0
-  MIN_POSSIBLE_GAS: float = 0.
-  MAX_POSSIBLE_GAS: float | None = None
+  MIN_POSSIBLE_GAS: int = 0.
+  MAX_POSSIBLE_GAS: int | None = None
 
   def test_gas_brake_limits_correct(self):
     self.assertIsNotNone(self.MAX_POSSIBLE_BRAKE)
